refactor solutions logic
